requirement.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
PyAudio==0.2.14          
 

fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
pydantic==2.7.1
 

python-dotenv==1.0.1